GeoLocation setup and implementation

+++ weather-app/src/app/pages/FavoritePage/index.tsx
@@ -6,6 +6,8 @@ import { TbTemperatureCelsius, TbTemperatureFahrenheit } from "react-icons/tb"
 import classnames from "classnames"
 import { useAppSelector } from "../../hooks"
 import DataVisualFavorites from "../../UI-components/DataVisualalizationFavorites"
+import Swal from "sweetalert2"
+import classNames from "classnames"
 
 export const daysOfWeek = [
   "Sunday",
@@ -19,129 +21,129 @@ export const daysOfWeek = [
 
 const FavoritePage = () => {
   const [favoriteList, setFavoriteList] = useState<any[]>(
-    // JSON.parse(localStorage.getItem('likedPlaces') as string) ||[]
-    [
-      {
-        LocalObservationDateTime: "2023-12-04T07:10:00-05:00",
-        EpochTime: 1701691800,
-        WeatherText: "Mostly cloudy",
-        WeatherIcon: 6,
-        HasPrecipitation: false,
-        PrecipitationType: null,
-        IsDayTime: true,
-        Temperature: {
-          Metric: {
-            Value: 21.6,
-            Unit: "C",
-            UnitType: 17,
-          },
-          Imperial: {
-            Value: 71,
-            Unit: "F",
-            UnitType: 18,
-          },
-        },
-        MobileLink:
-          "http://www.accuweather.com/en/ec/los-angeles/1239134/current-weather/1239134?lang=en-us",
-        Link: "http://www.accuweather.com/en/ec/los-angeles/1239134/current-weather/1239134?lang=en-us",
-      },
-      {
-        LocalObservationDateTime: "2023-12-04T07:10:00-05:00",
-        EpochTime: 1701691800,
-        WeatherText: "Mostly cloudy",
-        WeatherIcon: 6,
-        HasPrecipitation: false,
-        PrecipitationType: null,
-        IsDayTime: true,
-        Temperature: {
-          Metric: {
-            Value: 21.6,
-            Unit: "C",
-            UnitType: 17,
-          },
-          Imperial: {
-            Value: 71,
-            Unit: "F",
-            UnitType: 18,
-          },
-        },
-        MobileLink:
-          "httap://www.aaccuweather.com/en/ec/los-angeles/1239134/current-weather/1239134?lang=en-us",
-        Link: "htatp://waww.accuweather.com/en/ec/los-angeles/1239134/current-weather/1239134?lang=en-us",
-      },
-      {
-        LocalObservationDateTime: "2023-12-04T07:10:00-05:00",
-        EpochTime: 1701691800,
-        WeatherText: "Mostly cloudy",
-        WeatherIcon: 6,
-        HasPrecipitation: false,
-        PrecipitationType: null,
-        IsDayTime: true,
-        Temperature: {
-          Metric: {
-            Value: 21.6,
-            Unit: "C",
-            UnitType: 17,
-          },
-          Imperial: {
-            Value: 71,
-            Unit: "F",
-            UnitType: 18,
-          },
-        },
-        MobileLink:
-          "http://www.accuwaeather.com/en/ec/los-angeles/1239134/current-weather/1239134?lang=en-us",
-        Link: "http://www.acacuweather.com/en/ec/los-angeles/1239134/current-weather/1239134?lang=en-us",
-      },
-      {
-        LocalObservationDateTime: "2023-12-04T07:10:00-05:00",
-        EpochTime: 1701691800,
-        WeatherText: "Mostly cloudy",
-        WeatherIcon: 6,
-        HasPrecipitation: false,
-        PrecipitationType: null,
-        IsDayTime: true,
-        Temperature: {
-          Metric: {
-            Value: 21.6,
-            Unit: "C",
-            UnitType: 17,
-          },
-          Imperial: {
-            Value: 71,
-            Unit: "F",
-            UnitType: 18,
-          },
-        },
-        MobileLink:
-          "http://www.accuweasather.com/en/ec/los-angeles/1239134/current-weather/1239134?lang=en-us",
-        Link: "http://www.accsauweather.com/en/ec/los-angeles/1239134/current-weather/1239134?lang=en-us",
-      },
-      {
-        LocalObservationDateTime: "2023-12-04T07:10:00-05:00",
-        EpochTime: 1701691800,
-        WeatherText: "Mostly cloudy",
-        WeatherIcon: 6,
-        HasPrecipitation: false,
-        PrecipitationType: null,
-        IsDayTime: false,
-        Temperature: {
-          Metric: {
-            Value: 21.6,
-            Unit: "C",
-            UnitType: 17,
-          },
-          Imperial: {
-            Value: 71,
-            Unit: "F",
-            UnitType: 18,
-          },
-        },
-        MobileLink:
-          "http://www.accuweather.com/en/ec/slos-angeles/1239134/current-weather/1239134?lang=en-us",
-        Link: "http://www.accuweather.com/ens/ec/los-angeles/1239134/current-weather/1239134?lang=en-us",
-      },
-    ],
+    JSON.parse(localStorage.getItem('likedPlaces') as string) ||[]
+    // [
+    //   {
+    //     LocalObservationDateTime: "2023-12-04T07:10:00-05:00",
+    //     EpochTime: 1701691800,
+    //     WeatherText: "Mostly cloudy",
+    //     WeatherIcon: 6,
+    //     HasPrecipitation: false,
+    //     PrecipitationType: null,
+    //     IsDayTime: true,
+    //     Temperature: {
+    //       Metric: {
+    //         Value: 21.6,
+    //         Unit: "C",
+    //         UnitType: 17,
+    //       },
+    //       Imperial: {
+    //         Value: 71,
+    //         Unit: "F",
+    //         UnitType: 18,
+    //       },
+    //     },
+    //     MobileLink:
+    //       "http://www.accuweather.com/en/ec/los-angeles/1239134/current-weather/1239134?lang=en-us",
+    //     Link: "http://www.accuweather.com/en/ec/los-angeles/1239134/current-weather/1239134?lang=en-us",
+    //   },
+    //   {
+    //     LocalObservationDateTime: "2023-12-04T07:10:00-05:00",
+    //     EpochTime: 1701691800,
+    //     WeatherText: "Mostly cloudy",
+    //     WeatherIcon: 6,
+    //     HasPrecipitation: false,
+    //     PrecipitationType: null,
+    //     IsDayTime: true,
+    //     Temperature: {
+    //       Metric: {
+    //         Value: 21.6,
+    //         Unit: "C",
+    //         UnitType: 17,
+    //       },
+    //       Imperial: {
+    //         Value: 71,
+    //         Unit: "F",
+    //         UnitType: 18,
+    //       },
+    //     },
+    //     MobileLink:
+    //       "httap://www.aaccuweather.com/en/ec/los-angeles/1239134/current-weather/1239134?lang=en-us",
+    //     Link: "htatp://waww.accuweather.com/en/ec/los-angeles/1239134/current-weather/1239134?lang=en-us",
+    //   },
+    //   {
+    //     LocalObservationDateTime: "2023-12-04T07:10:00-05:00",
+    //     EpochTime: 1701691800,
+    //     WeatherText: "Mostly cloudy",
+    //     WeatherIcon: 6,
+    //     HasPrecipitation: false,
+    //     PrecipitationType: null,
+    //     IsDayTime: true,
+    //     Temperature: {
+    //       Metric: {
+    //         Value: 21.6,
+    //         Unit: "C",
+    //         UnitType: 17,
+    //       },
+    //       Imperial: {
+    //         Value: 71,
+    //         Unit: "F",
+    //         UnitType: 18,
+    //       },
+    //     },
+    //     MobileLink:
+    //       "http://www.accuwaeather.com/en/ec/los-angeles/1239134/current-weather/1239134?lang=en-us",
+    //     Link: "http://www.acacuweather.com/en/ec/los-angeles/1239134/current-weather/1239134?lang=en-us",
+    //   },
+    //   {
+    //     LocalObservationDateTime: "2023-12-04T07:10:00-05:00",
+    //     EpochTime: 1701691800,
+    //     WeatherText: "Mostly cloudy",
+    //     WeatherIcon: 6,
+    //     HasPrecipitation: false,
+    //     PrecipitationType: null,
+    //     IsDayTime: true,
+    //     Temperature: {
+    //       Metric: {
+    //         Value: 21.6,
+    //         Unit: "C",
+    //         UnitType: 17,
+    //       },
+    //       Imperial: {
+    //         Value: 71,
+    //         Unit: "F",
+    //         UnitType: 18,
+    //       },
+    //     },
+    //     MobileLink:
+    //       "http://www.accuweasather.com/en/ec/los-angeles/1239134/current-weather/1239134?lang=en-us",
+    //     Link: "http://www.accsauweather.com/en/ec/los-angeles/1239134/current-weather/1239134?lang=en-us",
+    //   },
+    //   {
+    //     LocalObservationDateTime: "2023-12-04T07:10:00-05:00",
+    //     EpochTime: 1701691800,
+    //     WeatherText: "Mostly cloudy",
+    //     WeatherIcon: 6,
+    //     HasPrecipitation: false,
+    //     PrecipitationType: null,
+    //     IsDayTime: false,
+    //     Temperature: {
+    //       Metric: {
+    //         Value: 21.6,
+    //         Unit: "C",
+    //         UnitType: 17,
+    //       },
+    //       Imperial: {
+    //         Value: 71,
+    //         Unit: "F",
+    //         UnitType: 18,
+    //       },
+    //     },
+    //     MobileLink:
+    //       "http://www.accuweather.com/en/ec/slos-angeles/1239134/current-weather/1239134?lang=en-us",
+    //     Link: "http://www.accuweather.com/ens/ec/los-angeles/1239134/current-weather/1239134?lang=en-us",
+    //   },
+    // ],
   )
   const toast = useRef<Toast | null>(null)
   const [isOnFahrenheit, setIsOnFahrenheit] = useState<boolean>(false)
@@ -149,11 +151,29 @@ const FavoritePage = () => {
   const theme = localStorage.getItem("theme")
   const handleRemoveLike = (place: any) => {
     const placeIdentifier = place.MobileLink
-    const updatedList = favoriteList.filter(
-      (p: any) => p.MobileLink !== placeIdentifier,
-    )
-    setFavoriteList(updatedList)
-    localStorage.setItem("likedPlaces", JSON.stringify(updatedList))
+    return Swal.fire({
+        title: `Are you sure you want to remove ${place.LocalizedName}?`,
+        // text: "",
+        icon: "warning",
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Confirm",
+        showCancelButton: true,
+        // cancelButtonText: "Cancel",
+    }).then((result) => {
+        if (!result.isConfirmed) {
+
+        } else {
+          const updatedList = favoriteList.filter((p: any) => p.MobileLink !== placeIdentifier);
+          setFavoriteList(updatedList);
+          localStorage.setItem('likedPlaces', JSON.stringify(updatedList));
+        }
+
+    });
+
+
+
+   
   }
 
   const handleTemperatureIndicatorSelect = () => {
@@ -190,21 +210,33 @@ const FavoritePage = () => {
                 {isOnFahrenheit ? (
                   <button
                     onClick={() => handleTemperatureIndicatorSelect()}
-                    className="text-2xl rounded-full mt-1 hover:text-gray-800 duration-300 ease-in hover:bg-gray-300 p-2 min-h-[2rem] transition-colors"
+                    className={classNames({
+                        "text-2xl rounded-full mt-1 duration-300 ease-in p-2 min-h-[2rem] transition-colors":true,
+                        "  hover:text-gray-800  hover:bg-gray-300":theme === "dark",
+                        "  hover:text-gray-200  hover:bg-gray-800":theme === "light",
+                    })}
                   >
                     <TbTemperatureCelsius className=" cursor-pointer" />
                   </button>
                 ) : (
                   <button
                     onClick={() => handleTemperatureIndicatorSelect()}
-                    className="text-2xl p-2 min-h-[2rem] duration-300 ease-in transition-colors"
+                    className={classNames({
+                        "text-2xl rounded-full mt-1 duration-300 ease-in p-2 min-h-[2rem] transition-colors":true,
+                        "  hover:text-gray-800  hover:bg-gray-300":theme === "dark",
+                        "  hover:text-gray-200  hover:bg-gray-800":theme === "light",
+                    })}
                   >
                     <TbTemperatureFahrenheit className=" cursor-pointer" />
                   </button>
                 )}
                 <button
                   onClick={() => handleRemoveLike(place)}
-                  className="text-2xl rounded-full mt-2 hover:bg-gray-300 duration-300 ease-in hover:text-gray-800 p-2 min-h-[2rem] transition-colors"
+                  className={classNames(
+                    {"text-2xl rounded-full mt-2 duration-300 ease-in  p-2 min-h-[2rem] transition-colors":true,
+                  "  hover:text-gray-800  hover:bg-gray-300":theme === "dark",
+                  "  hover:text-gray-200  hover:bg-gray-800":theme === "light",
+                })}
                 >
                   <IoHeartDislike className=" cursor-pointer" />
                 </button>
